colorspace: use bfloat16 for SSE2 linear-to-gamma LUT
On Skylake, processing 512 pixel array:

direction	cycles/sample	cycles/pxvector
     g->l	         2.35	           9.43
     l->g	         2.49	           9.97

+++ src/zimg/colorspace/x86/operation_impl_avx2.cpp
@@ -114,8 +114,7 @@ public:
 	{
 		EnsureSinglePrecision x87;
 
-		// Allocate an extra LUT entry so that indexing can be done by multipying by a power of 2.
-		for (unsigned long i = 0; i <= UINT16_MAX; ++i) {
+		for (size_t i = 0; i <= UINT16_MAX; ++i) {
 			uint16_t half = static_cast<uint16_t>(i);
 			float x = _mm_cvtss_f32(_mm_cvtph_ps(_mm_set1_epi16(half)));
 			m_lut[i] = func(x * prescale);
